src/api_gateway/requirements.txt
fastapi
asyncpg
passlib
python-jose
pydantic
uvicorn
python-multipart
grpcio 
grpcio-tools
email-validator